Added method for fetching the sample clock rate. Changed the signature of the Up converter write method to require that a 4 byte int pointer is used.
git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/pentek@7837 da89c675-a4d7-0310-8a15-9e7546503515

+++ test/dactest.cpp
@@ -49,7 +49,7 @@ int main(int argc, char** argv) {
 
 	// create the signal
 	unsigned int n = 100;
-	long IQ[n];
+	int32_t IQ[n];
 	for (unsigned int i = 0; i < n/4; i++) {
 		IQ[i]   = 0x7FFF << 16 | 0x7FFF;
 	}
